Add Daylight Savings rule & dialtone config
Note that Daylight Savings rules only apply if you are set $provisioner->DateTimeZone to a suitable object -
either the php 5.3 DateTimeZone object, or an instance of the object found in samples/tz.php.

The rule is worked out as part of the provisioning, so applications should not set it. Because of this, there is no new entry in template_data.xml

Possibly later on, some of the data generated here may be moved up into endpoint/base.php, so other phones can make use of it.